backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.8.2
pydantic-settings==2.4.0
pyyaml==6.0.2
httpx==0.27.2
aiohttp==3.10.5
anthropic==0.34.2
python-multipart==0.0.9
python-dotenv==1.0.1
pygments==2.18.0
tree-sitter==0.23.1
gitpython==3.1.43
requests==2.32.3
numpy==2.1.1
